new allowlist file infrastructure
